Internals

Index

GENeSYS_MOD.Emp_Sets

GENeSYS_MOD.InputClass

GENeSYS_MOD.Parameters

GENeSYS_MOD.Sets

GENeSYS_MOD.Settings

GENeSYS_MOD.SubsetsIni

GENeSYS_MOD.Switch

GENeSYS_MOD.Variable_Parameters

GENeSYS_MOD._registered_variables

GENeSYS_MOD.convert_jump_container_to_df

GENeSYS_MOD.create_daa

GENeSYS_MOD.create_daa_hourly

GENeSYS_MOD.create_daa_init

GENeSYS_MOD.create_df_hourly

GENeSYS_MOD.df_from_daa

GENeSYS_MOD.df_from_dd

GENeSYS_MOD.df_from_json

GENeSYS_MOD.genesysmod_bounds

GENeSYS_MOD.genesysmod_dataload

GENeSYS_MOD.genesysmod_dec

GENeSYS_MOD.genesysmod_emissionintensity

GENeSYS_MOD.genesysmod_employment

GENeSYS_MOD.genesysmod_equ

GENeSYS_MOD.genesysmod_levelizedcosts

GENeSYS_MOD.genesysmod_results

GENeSYS_MOD.genesysmod_results_raw

GENeSYS_MOD.genesysmod_settings

GENeSYS_MOD.genesysmod_variable_parameter

GENeSYS_MOD.make_subsets

GENeSYS_MOD.print_iis

GENeSYS_MOD.read_in_data

GENeSYS_MOD.read_time_series

GENeSYS_MOD.specified_demand_profile

GENeSYS_MOD.timeseries_reduction

Types

GENeSYS_MOD.Emp_Sets — Type

Sets necessary for the employment calculations. The sets elements may be different from the sets in the model.

Fields

Technology ::Union{Nothing,Array} TODO.

Year ::Union{Nothing,Array} TODO.

Region ::Union{Nothing,Array} TODO.

GENeSYS_MOD.InputClass — Type

InputClass - Abstract type for the inputs

GENeSYS_MOD.Parameters — Type

All the parameters read for a model run

Extended

Fields

StartYear ::Int64 First year of the study horizon.

YearSplit ::JuMP.Containers.DenseAxisArray Weighting factors of the each timeslice, i.e. how much of the whole year is represented by a given tiomeslice.

SpecifiedAnnualDemand ::JuMP.Containers.DenseAxisArray Total specified demand for a year.

SpecifiedDemandProfile ::JuMP.Containers.DenseAxisArray Annual fraction of energy-service or commodity demand that is required in each time slice. For each year, all the defined SpecifiedDemandProfile input values should sum up to 1.

RateOfDemand ::JuMP.Containers.DenseAxisArray

#TODO

Demand ::JuMP.Containers.DenseAxisArray

#TODO

CapacityToActivityUnit ::JuMP.Containers.DenseAxisArray Conversion factor relating the energy that would be produced when one unit of capacity is fully used in one year.

CapacityFactor ::JuMP.Containers.DenseAxisArray Capacity available per each TimeSlice expressed as a fraction of the total installed capacity, with values ranging from 0 to 1. It gives the possibility to account for forced outages or variable renewable generation.

AvailabilityFactor ::JuMP.Containers.DenseAxisArray Maximum time a technology can run in the whole year, as a fraction of the year ranging from 0 to 1. It gives the possibility to account for planned outages.

OperationalLife ::JuMP.Containers.DenseAxisArray Useful lifetime of a technology, expressed in years.

ResidualCapacity ::JuMP.Containers.DenseAxisArray

#TODO

InputActivityRatio ::JuMP.Containers.DenseAxisArray Rate of use of a fuel by a technology, as a ratio of the rate of activity. Used to express technology efficiencies.

OutputActivityRatio ::JuMP.Containers.DenseAxisArray Rate of fuel output from a technology, as a ratio of the rate of activity.

CapacityOfOneTechnologyUnit ::JuMP.Containers.DenseAxisArray

#TODO

TagDispatchableTechnology ::JuMP.Containers.DenseAxisArray

#TODO

BaseYearProduction ::JuMP.Containers.DenseAxisArray

#TODO

RegionalBaseYearProduction ::JuMP.Containers.DenseAxisArray

#TODO

RegionalCCSLimit ::JuMP.Containers.DenseAxisArray Total amount of storeable emissions in a certain region over the entire modelled period.

CapitalCost ::JuMP.Containers.DenseAxisArray Capital investment cost of a technology, per unit of capacity.

VariableCost ::JuMP.Containers.DenseAxisArray Cost of a technology for a given mode of operation (e.g., Variable O&M cost, fuel costs, etc.), per unit of activity.

FixedCost ::JuMP.Containers.DenseAxisArray Fixed O&M cost of a technology, per unit of capacity.

StorageLevelStart ::JuMP.Containers.DenseAxisArray Level of storage at the beginning of first modelled year, in units of activity.

StorageMaxChargeRate ::JuMP.Containers.DenseAxisArray Maximum charging rate for the storage, in units of activity per year.

StorageMaxDischargeRate ::JuMP.Containers.DenseAxisArray Maximum discharging rate for the storage, in units of activity per year.

MinStorageCharge ::JuMP.Containers.DenseAxisArray Sets a lower bound to the amount of energy stored, as a fraction of the maximum, with a number ranging between 0 and 1. The storage facility cannot be emptied below this level.

OperationalLifeStorage ::JuMP.Containers.DenseAxisArray Useful lifetime of the storage facility.

CapitalCostStorage ::JuMP.Containers.DenseAxisArray Binary parameter linking a technology to the storage facility it charges. It has value 0 if the technology and the storage facility are not linked, 1 if they are.

ResidualStorageCapacity ::JuMP.Containers.DenseAxisArray Binary parameter linking a storage facility to the technology it feeds. It has value 0 if the technology and the storage facility are not linked, 1 if they are.

TechnologyToStorage ::JuMP.Containers.DenseAxisArray Binary parameter linking a technology to the storage facility it charges. It has value 1 if the technology and the storage facility are linked, 0 otherwise.

TechnologyFromStorage ::JuMP.Containers.DenseAxisArray Binary parameter linking a storage facility to the technology it feeds. It has value 1 if the technology and the storage facility are linked, 0 otherwise.

StorageMaxCapacity ::JuMP.Containers.DenseAxisArray

#TODO

TotalAnnualMaxCapacity ::JuMP.Containers.DenseAxisArray Total maximum existing (residual plus cumulatively installed) capacity allowed for a technology in a specified year.

TotalAnnualMinCapacity ::JuMP.Containers.DenseAxisArray Total minimum existing (residual plus cumulatively installed) capacity allowed for a technology in a specified year.

TagTechnologyToSector ::JuMP.Containers.DenseAxisArray Links technologies to sectors.

AnnualSectoralEmissionLimit ::JuMP.Containers.DenseAxisArray Annual upper limit for a specific emission generated in a certain sector for the whole modelled region.

TotalAnnualMaxCapacityInvestment ::JuMP.Containers.DenseAxisArray Maximum capacity of a technology, expressed in power units.

TotalAnnualMinCapacityInvestment ::JuMP.Containers.DenseAxisArray Minimum capacity of a technology, expressed in power units.

TotalTechnologyAnnualActivityUpperLimit ::JuMP.Containers.DenseAxisArray Total maximum level of activity allowed for a technology in one year.

TotalTechnologyAnnualActivityLowerLimit ::JuMP.Containers.DenseAxisArray Total minimum level of activity allowed for a technology in one year.

TotalTechnologyModelPeriodActivityUpperLimit ::JuMP.Containers.DenseAxisArray Total maximum level of activity allowed for a technology in the entire modelled period.

TotalTechnologyModelPeriodActivityLowerLimit ::JuMP.Containers.DenseAxisArray Total minimum level of activity allowed for a technology in the entire modelled period.

ReserveMarginTagTechnology ::JuMP.Containers.DenseAxisArray Binary parameter tagging the technologies that are allowed to contribute to the reserve margin. It has value 1 for the technologies allowed, 0 otherwise.

ReserveMarginTagFuel ::JuMP.Containers.DenseAxisArray Binary parameter tagging the fuels to which the reserve margin applies. It has value 1 if the reserve margin applies to the fuel, 0 otherwise.

ReserveMargin ::JuMP.Containers.DenseAxisArray Minimum level of the reserve margin required to be provided for all the tagged commodities, by the tagged technologies. If no reserve margin is required, the parameter will have value 1; if, for instance, 20% reserve margin is required, the parameter will have value 1.2.

RETagTechnology ::JuMP.Containers.DenseAxisArray

#TODO

RETagFuel ::JuMP.Containers.DenseAxisArray

#TODO

REMinProductionTarget ::JuMP.Containers.DenseAxisArray

#TODO

EmissionActivityRatio ::JuMP.Containers.DenseAxisArray Emission factor of a technology per unit of activity, per mode of operation.

EmissionContentPerFuel ::JuMP.Containers.DenseAxisArray Defines the emission contents per fuel.

EmissionsPenalty ::JuMP.Containers.DenseAxisArray Monetary penalty per unit of emission.

EmissionsPenaltyTagTechnology ::JuMP.Containers.DenseAxisArray Activates or deactivates emission penalties for specific technologies.

AnnualExogenousEmission ::JuMP.Containers.DenseAxisArray Additional annual emissions, on top of those computed endogenously by the model.

AnnualEmissionLimit ::JuMP.Containers.DenseAxisArray Annual upper limit for a specific emission generated in the whole modelled region.

RegionalAnnualEmissionLimit ::JuMP.Containers.DenseAxisArray Annual upper limit for a specific emission generated in a certain modelled region.

ModelPeriodExogenousEmission ::JuMP.Containers.DenseAxisArray Additional emissions over the entire modelled period, on top of those computed endogenously by the model.

ModelPeriodEmissionLimit ::JuMP.Containers.DenseAxisArray Total model period upper limit for a specific emission generated in the whole modelled region.

RegionalModelPeriodEmissionLimit ::JuMP.Containers.DenseAxisArray Total model period upper limit for a specific emission generated in a certain modelled region.

CurtailmentCostFactor ::JuMP.Containers.DenseAxisArray Costs per curtailed unit of activity for certain fuels and years.

TradeRoute ::JuMP.Containers.DenseAxisArray Sets the distance in km from one region to another. Also controls the ability to trade on fuel from a region to another.

TradeCosts ::JuMP.Containers.DenseAxisArray Costs for trading one unit of energy from one region to another.

TradeLossFactor ::JuMP.Containers.DenseAxisArray

#TODO

TradeRouteInstalledCapacity ::JuMP.Containers.DenseAxisArray

#TODO

TradeLossBetweenRegions ::JuMP.Containers.DenseAxisArray Percentage loss of traded fuel from one region to another. Used to model losses in power transmission networks.

AdditionalTradeCapacity ::JuMP.Containers.DenseAxisArray

#TODO

TradeCapacity ::JuMP.Containers.DenseAxisArray Initial capacity for trading fuels from one region to another.

TradeCapacityGrowthCosts ::JuMP.Containers.DenseAxisArray Costs for adding one unit of additional trade capacity per km from one region to another.

GrowthRateTradeCapacity ::JuMP.Containers.DenseAxisArray Upper limit for adding additional trade capacities. Given as maximal percentage increase of installed capacity.

SelfSufficiency ::JuMP.Containers.DenseAxisArray Lower bound that limits the imports of fuels in as specific year and region.

RampingUpFactor ::Union{Nothing,JuMP.Containers.DenseAxisArray} Defines how much of the built capacity can be activated each timeslice.

RampingDownFactor ::Union{Nothing,JuMP.Containers.DenseAxisArray} Defines how much of the built capacity can be deactivated each timeslice.

ProductionChangeCost ::Union{Nothing,JuMP.Containers.DenseAxisArray} Cost per unit of activated or deactivated capacity per timeslice.

MinActiveProductionPerTimeslice ::Union{Nothing,JuMP.Containers.DenseAxisArray} Minimum fuel production from specific technologies in a certain timeslice. Represents minimum active capacity requirements.

ModalSplitByFuelAndModalType ::JuMP.Containers.DenseAxisArray Lower bound of production of certain fuels by specific modal types.

TagTechnologyToModalType ::JuMP.Containers.DenseAxisArray Links technology production by mode of operation to modal stype.

EFactorConstruction ::Union{Nothing,JuMP.Containers.DenseAxisArray}

#TODO

EFactorOM ::Union{Nothing,JuMP.Containers.DenseAxisArray}

#TODO

EFactorManufacturing ::Union{Nothing,JuMP.Containers.DenseAxisArray}

#TODO

EFactorFuelSupply ::Union{Nothing,JuMP.Containers.DenseAxisArray}

#TODO

EFactorCoalJobs ::Union{Nothing,JuMP.Containers.DenseAxisArray}

#TODO

CoalSupply ::Union{Nothing,JuMP.Containers.DenseAxisArray}

#TODO

CoalDigging ::Union{Nothing,JuMP.Containers.DenseAxisArray}

#TODO

RegionalAdjustmentFactor ::Union{Nothing,JuMP.Containers.DenseAxisArray}

#TODO

LocalManufacturingFactor ::Union{Nothing,JuMP.Containers.DenseAxisArray}

#TODO

DeclineRate ::Union{Nothing,JuMP.Containers.DenseAxisArray}

#TODO

x_peakingDemand ::Union{Nothing,JuMP.Containers.DenseAxisArray}

#TODO

TagDemandFuelToSector ::JuMP.Containers.DenseAxisArray

#TODO

TagElectricTechnology ::JuMP.Containers.DenseAxisArray

#TODO

GENeSYS_MOD.Sets — Type

Sets used for the model run

Fields

Timeslice_full ::Array TODO.

DailyTimeBracket ::Array TODO.

Year_full ::Array TODO.

Emission ::Array TODO.

Technology ::Array TODO.

Fuel ::Array TODO.

Year ::Array TODO.

Timeslice ::Array TODO.

Mode_of_operation ::Array TODO.

Region_full ::Array TODO.

Season ::Array TODO.

Daytype ::Array TODO.

Storage ::Array TODO.

ModalType ::Array TODO.

Sector ::Array TODO.

GENeSYS_MOD.Settings — Type

Model settings necessary for running the model

Fields

DepreciationMethod ::JuMP.Containers.DenseAxisArray Parameter defining the type of depreciation to be applied in each region. It has value 1 for sinking fund depreciation, value 2 for straight-line depreciation.

GeneralDiscountRate ::JuMP.Containers.DenseAxisArray TODO.

TechnologyDiscountRate ::JuMP.Containers.DenseAxisArray TODO.

SocialDiscountRate ::JuMP.Containers.DenseAxisArray TODO.

InvestmentLimit TODO.

NewRESCapacity ::Float64 TODO.

ProductionGrowthLimit ::JuMP.Containers.DenseAxisArray TODO.

StorageLimitOffset ::Float64 TODO.

Trajectory2020UpperLimit ::Float64 TODO.

Trajectory2020LowerLimit ::Float64 TODO.

BaseYearSlack TODO.

PhaseIn ::Dict TODO.

PhaseOut ::Dict TODO.

GENeSYS_MOD.SubsetsIni — Type

Subsets of the Sets used to define in more detail the characteristics of some members the sets.

Fields

Solar TODO.

Wind TODO.

Renewables TODO.

CCS TODO.

Transformation TODO.

RenewableTransformation TODO.

FossilFuelGeneration TODO.

FossilFuels TODO.

FossilPower TODO.

CHPs TODO.

RenewableTransport TODO.

Transport TODO.

Passenger TODO.

Freight TODO.

TransportFuels TODO.

ImportTechnology TODO.

Heat TODO.

PowerSupply TODO.

PowerBiomass TODO.

Coal TODO.

Lignite TODO.

Gas TODO.

StorageDummies TODO.

SectorCoupling TODO.

HeatFuels TODO.

ModalGroups TODO.

PhaseInSet TODO.

PhaseOutSet TODO.

HeatSlowRamper TODO.

HeatQuickRamper TODO.

Hydro TODO.

Geothermal TODO.

Onshore TODO.

Offshore TODO.

SolarUtility TODO.

Oil TODO.

HeatLowRes TODO.

HeatLowInd TODO.

HeatMedInd TODO.

HeatHighInd TODO.

Biomass TODO.

Households TODO.

Companies TODO.

HydrogenTechnologies TODO.

DummyTechnology TODO.

GENeSYS_MOD.Switch — Type

Switches used to define a model run.

The switches corresponds to various elements and define for instance the input files and folders, as well as the inclusion or not of various features.

Fields

StartYear :: Int16 TODO.

switch_only_load_gdx ::Int8 TODO.

switch_test_data_load ::Int8 TODO.

solver TODO.

DNLPsolver TODO.

model_region ::String TODO.

data_base_region ::String TODO.

data_file ::String TODO.

timeseries_data_file ::String TODO.

threads ::Int16 TODO.

emissionPathway ::String TODO.

emissionScenario ::String TODO.

socialdiscountrate ::Float64 TODO.

inputdir ::String TODO.

tempdir ::String TODO.

resultdir ::String TODO.

switch_infeasibility_tech :: Int8 TODO.

switch_investLimit ::Int16 TODO.

switch_ccs ::Int16 TODO.

switch_ramping ::Int16 TODO.

switch_weighted_emissions ::Int16 TODO.

switch_intertemporal ::Int16 TODO.

switch_base_year_bounds ::Int16 TODO.

switch_peaking_capacity ::Int16 TODO.

set_peaking_slack ::Float16 TODO.

set_peaking_minrun_share ::Float16 TODO.

set_peaking_res_cf ::Float16 TODO.

set_peaking_startyear ::Int16 TODO.

switch_peaking_with_storages ::Int16 TODO.

switch_peaking_with_trade ::Int16 TODO.

switch_peaking_minrun ::Int16 TODO.

switch_employment_calculation ::Int16 TODO.

switch_endogenous_employment ::Int16 TODO.

employment_data_file ::String TODO.

switch_dispatch ::Int8 TODO.

hourly_data_file ::String TODO.

elmod_nthhour ::Int16 TODO.

elmod_starthour ::Int16 TODO.

elmod_dunkelflaute ::Int16 TODO.

elmod_daystep ::Int16 TODO.

elmod_hourstep ::Int16 TODO.

switch_raw_results ::Int8 TODO.

switch_processed_results ::Int8 TODO.

write_reduced_timeserie ::Int8 TODO.

GENeSYS_MOD.Variable_Parameters — Type

Intermediary variables calculated after the model run

The intermediary variables are used to ex-post aggregate the activity, demand and use by time, technology, mod of operation and/or region.

Methods

GENeSYS_MOD._registered_variables — Method

GENeSYS_MOD.convert_jump_container_to_df — Method

Returns a DataFrame with the values of the variables from the JuMP container var. The column names of the DataFrame can be specified for the indexing columns in dim_names, and the name of the data value column by a Symbol value_col e.g. :Value

GENeSYS_MOD.create_daa — Function

GENeSYS_MOD.create_daa_hourly — Method

GENeSYS_MOD.create_daa_init — Function

GENeSYS_MOD.create_df_hourly — Method

GENeSYS_MOD.df_from_daa — Method

GENeSYS_MOD.df_from_dd — Method

GENeSYS_MOD.df_from_json — Method

GENeSYS_MOD.genesysmod_bounds — Method

GENeSYS_MOD.genesysmod_dataload — Method

GENeSYS_MOD.genesysmod_dec — Method

GENeSYS_MOD.genesysmod_emissionintensity — Method

GENeSYS_MOD.genesysmod_employment — Method

GENeSYS_MOD.genesysmod_equ — Method

GENeSYS_MOD.genesysmod_levelizedcosts — Method

GENeSYS_MOD.genesysmod_results — Method

GENeSYS_MOD.genesysmod_results_raw — Method

GENeSYS_MOD.genesysmod_settings — Method

GENeSYS_MOD.genesysmod_variable_parameter — Method

GENeSYS_MOD.make_subsets — Method

GENeSYS_MOD.print_iis — Method

GENeSYS_MOD.read_in_data — Method

GENeSYS_MOD.read_time_series — Method

GENeSYS_MOD.specified_demand_profile — Function

GENeSYS_MOD.timeseries_reduction — Method