Internals

Index

GENeSYS_MOD.AbstractAggregationMode

GENeSYS_MOD.AbstractDispatchType

GENeSYS_MOD.AbstractInfeasibilityTechs

GENeSYS_MOD.Emp_Sets

GENeSYS_MOD.InputClass

GENeSYS_MOD.Maps

GENeSYS_MOD.Mean

GENeSYS_MOD.NoDispatch

GENeSYS_MOD.Parameters

GENeSYS_MOD.RawResultType

GENeSYS_MOD.Sets

GENeSYS_MOD.Settings

GENeSYS_MOD.SubsetsIni

GENeSYS_MOD.Sum

GENeSYS_MOD.Switch

GENeSYS_MOD.Tags

GENeSYS_MOD.Variable_Parameters

GENeSYS_MOD._registered_variables

GENeSYS_MOD.convert_jump_container_to_df

GENeSYS_MOD.create_daa

GENeSYS_MOD.create_daa_hourly

GENeSYS_MOD.create_daa_init

GENeSYS_MOD.create_df_hourly

GENeSYS_MOD.genesysmod_bounds

GENeSYS_MOD.genesysmod_dataload

GENeSYS_MOD.genesysmod_dec

GENeSYS_MOD.genesysmod_emissionintensity

GENeSYS_MOD.genesysmod_employment

GENeSYS_MOD.genesysmod_equ

GENeSYS_MOD.genesysmod_levelizedcosts

GENeSYS_MOD.genesysmod_results

GENeSYS_MOD.genesysmod_results_raw

GENeSYS_MOD.genesysmod_settings

GENeSYS_MOD.genesysmod_variable_parameter

GENeSYS_MOD.make_mapping

GENeSYS_MOD.print_iis

GENeSYS_MOD.simplify_iis

GENeSYS_MOD.split_by_substrings

GENeSYS_MOD.timeseries_reduction!

Types

GENeSYS_MOD.AbstractAggregationMode — Type

AbstractAggregationMode - Abstract type for the methode of aggregation used in the dispatch.

GENeSYS_MOD.AbstractDispatchType — Type

AbstractDispatchType - Type of dispatch to be used in the dispatch run.

GENeSYS_MOD.AbstractInfeasibilityTechs — Type

AbstractInfeasibilityTechs - abstract type for switchinfeasibilitytech.

GENeSYS_MOD.Emp_Sets — Type

Sets necessary for the employment calculations. The set elements may be different from the sets in the model.

GENeSYS_MOD.InputClass — Type

InputClass - Abstract type for the inputs

GENeSYS_MOD.Maps — Type

Structure containing the mappings of sets combinations

GENeSYS_MOD.Mean — Type

Mean - aggregation method taking the mean of the values.

GENeSYS_MOD.NoDispatch — Type

NoDispatch - No dispatch to be run. The type is used for investment runs.

GENeSYS_MOD.Parameters — Type

All the parameters read for a model run

Extended

Fields

StartYear ::Int64 First year of the study horizon.

YearSplit ::JuMP.Containers.DenseAxisArray Weighting factors of the each timeslice, i.e. how much of the whole year is represented by a given tiomeslice.

SpecifiedAnnualDemand ::JuMP.Containers.DenseAxisArray Total specified demand for a year.

SpecifiedDemandProfile ::JuMP.Containers.DenseAxisArray Annual fraction of energy-service or commodity demand that is required in each time slice. For each year, all the defined SpecifiedDemandProfile input values should sum up to 1.

RateOfDemand ::JuMP.Containers.DenseAxisArray Rate of Demand in each timeslice.

Demand ::JuMP.Containers.DenseAxisArray Amount of demand in each timeslice.

CapacityToActivityUnit ::JuMP.Containers.DenseAxisArray Conversion factor relating the energy that would be produced when one unit of capacity is fully used in one year.

CapacityFactor ::JuMP.Containers.DenseAxisArray Capacity available per each TimeSlice expressed as a fraction of the total installed capacity, with values ranging from 0 to 1. It gives the possibility to account for forced outages or variable renewable generation.

AvailabilityFactor ::JuMP.Containers.DenseAxisArray Maximum time a technology can run in the whole year, as a fraction of the year ranging from 0 to 1. It gives the possibility to account for planned outages.

OperationalLife ::JuMP.Containers.DenseAxisArray Useful lifetime of a technology, expressed in years.

ResidualCapacity ::JuMP.Containers.DenseAxisArray Amount of capacity from the capacity

existing in the system in the start year to remain in the given year.

InputActivityRatio ::JuMP.Containers.DenseAxisArray Rate of use of a fuel by a technology, as a ratio of the rate of activity. Used to express technology efficiencies.

OutputActivityRatio ::JuMP.Containers.DenseAxisArray Rate of fuel output from a technology, as a ratio of the rate of activity.

TagDispatchableTechnology ::JuMP.Containers.DenseAxisArray Tag defining if a technology

can be dispatched.

RegionalBaseYearProduction ::JuMP.Containers.DenseAxisArray Amount of energy produced by

technologies in the start year. Used if switchbaseyear_bounds is set to 1.

RegionalCCSLimit ::JuMP.Containers.DenseAxisArray Total amount of storeable emissions in a certain region over the entire modelled period.

CapitalCost ::JuMP.Containers.DenseAxisArray Capital investment cost of a technology, per unit of capacity.

VariableCost ::JuMP.Containers.DenseAxisArray Cost of a technology for a given mode of operation (e.g., Variable O&M cost, fuel costs, etc.), per unit of activity.

FixedCost ::JuMP.Containers.DenseAxisArray Fixed O&M cost of a technology, per unit of capacity.

StorageLevelStart ::JuMP.Containers.DenseAxisArray Level of storage at the beginning of first modelled year, in units of activity.

MinStorageCharge ::JuMP.Containers.DenseAxisArray Sets a lower bound to the amount of energy stored, as a fraction of the maximum, with a number ranging between 0 and 1. The storage facility cannot be emptied below this level.

OperationalLifeStorage ::JuMP.Containers.DenseAxisArray Useful lifetime of the storage facility.

CapitalCostStorage ::JuMP.Containers.DenseAxisArray Binary parameter linking a technology to the storage facility it charges. It has value 0 if the technology and the storage facility are not linked, 1 if they are.

ResidualStorageCapacity ::JuMP.Containers.DenseAxisArray Binary parameter linking a storage facility to the technology it feeds. It has value 0 if the technology and the storage facility are not linked, 1 if they are.

TechnologyToStorage ::JuMP.Containers.DenseAxisArray Binary parameter linking a technology to the storage facility it charges. It has value 1 if the technology and the storage facility are linked, 0 otherwise.

TechnologyFromStorage ::JuMP.Containers.DenseAxisArray Binary parameter linking a storage facility to the technology it feeds. It has value 1 if the technology and the storage facility are linked, 0 otherwise.

StorageMaxCapacity ::JuMP.Containers.DenseAxisArray Maximum storage capacity.

TotalAnnualMaxCapacity ::JuMP.Containers.DenseAxisArray Total maximum existing (residual plus cumulatively installed) capacity allowed for a technology in a specified year.

TotalAnnualMinCapacity ::JuMP.Containers.DenseAxisArray Total minimum existing (residual plus cumulatively installed) capacity allowed for a technology in a specified year.

TagTechnologyToSector ::JuMP.Containers.DenseAxisArray Links technologies to sectors.

AnnualSectoralEmissionLimit ::JuMP.Containers.DenseAxisArray Annual upper limit for a specific emission generated in a certain sector for the whole modelled region.

TotalAnnualMaxCapacityInvestment ::JuMP.Containers.DenseAxisArray Maximum capacity of a technology, expressed in power units.

TotalAnnualMinCapacityInvestment ::JuMP.Containers.DenseAxisArray Minimum capacity of a technology, expressed in power units.

TotalTechnologyAnnualActivityUpperLimit ::JuMP.Containers.DenseAxisArray Total maximum level of activity allowed for a technology in one year.

TotalTechnologyAnnualActivityLowerLimit ::JuMP.Containers.DenseAxisArray Total minimum level of activity allowed for a technology in one year.

TotalTechnologyModelPeriodActivityUpperLimit ::JuMP.Containers.DenseAxisArray Total maximum level of activity allowed for a technology in the entire modelled period.

TotalTechnologyModelPeriodActivityLowerLimit ::JuMP.Containers.DenseAxisArray Total minimum level of activity allowed for a technology in the entire modelled period.

ReserveMarginTagTechnology ::JuMP.Containers.DenseAxisArray Binary parameter tagging the technologies that are allowed to contribute to the reserve margin. It has value 1 for the technologies allowed, 0 otherwise.

ReserveMarginTagFuel ::JuMP.Containers.DenseAxisArray Binary parameter tagging the fuels to which the reserve margin applies. It has value 1 if the reserve margin applies to the fuel, 0 otherwise.

ReserveMargin ::JuMP.Containers.DenseAxisArray Minimum level of the reserve margin required to be provided for all the tagged commodities, by the tagged technologies. If no reserve margin is required, the parameter will have value 1; if, for instance, 20% reserve margin is required, the parameter will have value 1.2.

RETagTechnology ::JuMP.Containers.DenseAxisArray Tag to identify renewable technologies.

RETagFuel ::JuMP.Containers.DenseAxisArray Tag to identify renewable fuels.

REMinProductionTarget ::JuMP.Containers.DenseAxisArray Minimum production from renewable technologies.

EmissionActivityRatio ::JuMP.Containers.DenseAxisArray Emission factor of a technology per unit of activity, per mode of operation.

EmissionContentPerFuel ::JuMP.Containers.DenseAxisArray Defines the emission contents per fuel.

EmissionsPenalty ::JuMP.Containers.DenseAxisArray Monetary penalty per unit of emission.

EmissionsPenaltyTagTechnology ::JuMP.Containers.DenseAxisArray Activates or deactivates emission penalties for specific technologies.

AnnualExogenousEmission ::JuMP.Containers.DenseAxisArray Additional annual emissions, on top of those computed endogenously by the model.

AnnualEmissionLimit ::JuMP.Containers.DenseAxisArray Annual upper limit for a specific emission generated in the whole modelled region.

RegionalAnnualEmissionLimit ::JuMP.Containers.DenseAxisArray Annual upper limit for a specific emission generated in a certain modelled region.

ModelPeriodExogenousEmission ::JuMP.Containers.DenseAxisArray Additional emissions over the entire modelled period, on top of those computed endogenously by the model.

ModelPeriodEmissionLimit ::JuMP.Containers.DenseAxisArray Total model period upper limit for a specific emission generated in the whole modelled region.

RegionalModelPeriodEmissionLimit ::JuMP.Containers.DenseAxisArray Total model period upper limit for a specific emission generated in a certain modelled region.

CurtailmentCostFactor ::JuMP.Containers.DenseAxisArray Costs per curtailed unit of activity for certain fuels and years.

TradeRoute ::JuMP.Containers.DenseAxisArray Sets the distance in km from one region to another. Also controls the ability to trade on fuel from a region to another.

TradeCosts ::JuMP.Containers.DenseAxisArray Costs for trading one unit of energy from one region to another.

TradeLossFactor ::JuMP.Containers.DenseAxisArray Factor for the amount of losses per kilometer of a given fuel

TradeRouteInstalledCapacity ::JuMP.Containers.DenseAxisArray Installed transmission capacity between nodes.

TradeLossBetweenRegions ::JuMP.Containers.DenseAxisArray Percentage loss of traded fuel from one region to another. Used to model losses in power transmission networks.

CommissionedTradeCapacity ::JuMP.Containers.DenseAxisArray Transmission line already commissioned.

TradeCapacity ::JuMP.Containers.DenseAxisArray Initial capacity for trading fuels from one region to another.

TradeCapacityGrowthCosts ::JuMP.Containers.DenseAxisArray Costs for adding one unit of additional trade capacity per km from one region to another.

GrowthRateTradeCapacity ::JuMP.Containers.DenseAxisArray Upper limit for adding additional trade capacities. Given as maximal percentage increase of installed capacity.

SelfSufficiency ::JuMP.Containers.DenseAxisArray Lower bound that limits the imports of fuels in as specific year and region.

RampingUpFactor ::Union{Nothing,JuMP.Containers.DenseAxisArray} Defines how much of the built capacity can be activated each timeslice.

RampingDownFactor ::Union{Nothing,JuMP.Containers.DenseAxisArray} Defines how much of the built capacity can be deactivated each timeslice.

ProductionChangeCost ::Union{Nothing,JuMP.Containers.DenseAxisArray} Cost per unit of activated or deactivated capacity per timeslice.

MinActiveProductionPerTimeslice ::Union{Nothing,JuMP.Containers.DenseAxisArray} Minimum fuel production from specific technologies in a certain timeslice. Represents minimum active capacity requirements.

ModalSplitByFuelAndModalType ::JuMP.Containers.DenseAxisArray Lower bound of production of certain fuels by specific modal types.

TagTechnologyToModalType ::JuMP.Containers.DenseAxisArray Links technology production by mode of operation to modal stype.

EFactorConstruction ::Union{Nothing,JuMP.Containers.DenseAxisArray}

TODO

EFactorOM ::Union{Nothing,JuMP.Containers.DenseAxisArray}

TODO

EFactorManufacturing ::Union{Nothing,JuMP.Containers.DenseAxisArray}

TODO

EFactorFuelSupply ::Union{Nothing,JuMP.Containers.DenseAxisArray}

TODO

EFactorCoalJobs ::Union{Nothing,JuMP.Containers.DenseAxisArray}

TODO

CoalSupply ::Union{Nothing,JuMP.Containers.DenseAxisArray}

TODO

CoalDigging ::Union{Nothing,JuMP.Containers.DenseAxisArray}

TODO

RegionalAdjustmentFactor ::Union{Nothing,JuMP.Containers.DenseAxisArray}

TODO

LocalManufacturingFactor ::Union{Nothing,JuMP.Containers.DenseAxisArray}

TODO

DeclineRate ::Union{Nothing,JuMP.Containers.DenseAxisArray}

TODO

x_peakingDemand ::Union{Nothing,JuMP.Containers.DenseAxisArray} Peak demand in the original timeseries. Used for the peaking constraints.

TagDemandFuelToSector ::JuMP.Containers.DenseAxisArray Tag to link fuels to sectors.

TagElectricTechnology ::JuMP.Containers.DenseAxisArray Indicate if a technology is considered to be "direct electrification".

GENeSYS_MOD.RawResultType — Type

RawResultType - abstract type for the format of raw results desired.

GENeSYS_MOD.Sets — Type

Sets used for the model run

Fields

Timeslice_full ::Array Represents all timeslices within a Year. This usally means 8760 elements.

Emission ::Array Represents potential emissions that can be derived by the operation of certain Technologies. Typically this includes atmospheric emissions, such as CO2.

Technology ::Array Represents the main elements of the energy system that produce, convert, or transform energy (carriers) and their proxies. Technologies can represent specific individual technology options, such as a "Natural Gas CCGT". They can also represent abstracted or aggregated collection of technologies used for accounting purposes (e.g., stock of cars).

Fuel ::Array Represents energy carriers, energy services, or proxies that are consumed, produced, or transformed by Technologies. These can represent individual energy carriers, aggregated groups, or artificial commodities, required by the analysis to be carried out.

Year ::Array Represents the time-frame of the model. This set contains all years to be considered in the corresponding analysis.

Timeslice ::Array Represents the temporal resolution within a Year of the analysis. This set contains (reduced) consecutive hourly time-series. Each Year has the same amount of timeslices assigned.

Mode_of_operation ::Array Represents the different modes in which a technologies can be operated by. A technology can have different inputs and/or outputs for each mode of operation to represent fuel-switching. E.g., a CHP can produce electricity in one mode and heat in another one.

Region_full ::Array Represents the regional scope of the model. This set can usually contains aggregated global regions, individual countries, or subcountry regions and states.

Storage ::Array Represents storage facilities in the model. Storages can store FUELS and are linked to specific TECHNOLOGIES.

ModalType ::Array Represents a modal type (e.g., rail-transportation) used in the transportation sector of the model. These are used to control modal shifting to a certain degree.

Sector ::Array Represents different sectors in the energy system. Used for aggregation and accounting purposes.

GENeSYS_MOD.Settings — Type

Model settings necessary for running the model

Fields

DepreciationMethod ::JuMP.Containers.DenseAxisArray Parameter defining the type of depreciation to be applied in each region. It has value 1 for sinking fund depreciation, value 2 for straight-line depreciation.

GeneralDiscountRate ::JuMP.Containers.DenseAxisArray TODO.

TechnologyDiscountRate ::JuMP.Containers.DenseAxisArray TODO.

SocialDiscountRate ::JuMP.Containers.DenseAxisArray TODO.

InvestmentLimit TODO.

NewRESCapacity ::Float64 TODO.

ProductionGrowthLimit ::JuMP.Containers.DenseAxisArray This parameter controls the maximal increase between two years of a specific fuel production from renewable energy sources.

StorageLimitOffset ::Float64 TODO.

Trajectory2020UpperLimit ::Float64 TODO.

Trajectory2020LowerLimit ::Float64 TODO.

BaseYearSlack The allowed slack from the defined RegionalBaseYearProduction. Value between 0 and 1. Used only if switchbaseyear_bounds is set to 1.

PhaseIn ::Dict TODO.

PhaseOut ::Dict TODO.

StorageLevelYearStartUpperLimit ::Float64 TODO.

StorageLevelYearStartLowerLimit ::Float64 TODO.

GENeSYS_MOD.SubsetsIni — Type

Subsets of the Sets used to define in more detail the characteristics of some members the sets.

Fields

Solar Solar technologies.

Wind Wind technologies.

Renewables Non-fossil technologies

CCS Technologies linked to varbon capture and storage (CCS).

Transformation TODO.

RenewableTransformation TODO.

FossilFuelGeneration Sources of fossil fuel.

FossilFuels

FossilPower Technology using fossil fuels to produce power.

CHPs Combined heat and power technologies.

RenewableTransport Non-fossil transportation technologies.

Transport Transportation technologies.

Passenger Transportation technologies for passengers.

Freight Transportation technologies for freight.

TransportFuels Elements of the Fuel set corresponding to transport demand.

ImportTechnology Technologies corresponding to the import of resources.

Heat Technologies producing heat.

PowerSupply Technologies producing power.

PowerBiomass Technologies producing powerfrom biomass.

Coal Technologies consuming coal.

Lignite Technologies consuming lignite.

Gas Technologies consuming gas.

StorageDummies Technologies corresponding to the defined storages.

SectorCoupling TODO.

HeatFuels Elements of the Fuel set corresponding to heat demand.

ModalGroups Overarching groups of modal types.

PhaseInSet Technologies that can be subject to a phase in constraint, i.e. for which the production in a given year must be at least as much as the production from the previous year times a factor combining the evolution of demand and user input (<1).

PhaseOutSet Technologies that can be subject to a phase out constraint, i.e. for which the production in a given year must be at most equal to the production from the previous year times a factor combining the evolution of demand and user factor (>1).

HeatSlowRamper Heating technologies with long ramp up time.

HeatQuickRamper Heating technologies with short ramp up time.

Hydro Hydropower technologies.

Geothermal Geothermal technologies.

Onshore Onshore wind technologies.

Offshore Offshore wind technologies.

SolarUtility Utility scale solar technologies.

Oil Technologies consuming oil.

HeatLowRes Technologies that can produce the fuel Heat Low Residential.

HeatLowInd Technologies that can produce the fuel Heat Low Industrial.

HeatMedInd Technologies that can produce the fuel Heat Medium Industrial.

HeatHighInd Technologies that can produce the fuel Heat High Industrial.

Biomass Technologies producing the fuel biomass.

Households Technologies at the residential scale.

Companies Technoologies not at the residential scale.

HydrogenTechnologies Hydrogen technologies.

DummyTechnology Dummy technoliges used to debug when getting infeasible model.

GENeSYS_MOD.Sum — Type

Sum - aggregation method summing all values

GENeSYS_MOD.Switch — Type

Switches used to define a model run.

The switches corresponds to various elements and define for instance the input files and folders, as well as the inclusion or not of various features.

Fields

StartYear :: Int16 First year of the study horizon.

solver Solver to be used to solve the LP. The corresponding package must be called.

For instance: using Gurobi, then the solver is Gurobi.Optimizer.

DNLPsolver Solver used for the time reduction algorithm. The recommended solver

is Ipopt as it is open but other commercial solvers with a julia integration can be used.

model_region ::String Name of the modelled region. It will be used in naming files.

data_base_region ::String Default region of the model. The missing data will be copied

from the default region.

data_file ::String Path to the main input data file.

hourly_data_file ::String Path to the input data file containing the timeseries.

threads ::Int16 Number of threads to use for solving the model. Default is 4. To

automatically use th emaximum available number of threads, use the value 0.

emissionPathway ::String Name of the emission pathway. Used in naming files.

emissionScenario ::String Name of the emission scenario. Used in naming files.

socialdiscountrate ::Float64 Sets the value of the setting social discount rate.

inputdir ::String Directory containing the input files.

resultdir ::String Directory where the results files will be written.

switch_infeasibility_tech :: Int8 Switch used to include the infeasibility

technologies in the model. These technologies are used to debug an infeasible model and allow the model to run feasible by relaxing the problem with slack production technologies.

switch_investLimit ::Int16 Used to enable Investment limits. This activates phase in

constraints for renewable technologies, phase out constraints for fossil technologies. It also activates a constraint smoothing the investment and to prevent large investment in a single period. It also activates a constraint limiting the investment in renewable each year to a percentage of the total technical potential.

switch_ccs ::Int16 Used to enable CCS technologies.

switch_ramping ::Int16 Used to enable ramping constraints.

switch_weighted_emissions ::Int16 TODO.

switch_intertemporal ::Int16 TODO.

switch_base_year_bounds ::Int16 Used to enable base year bounds. This enforces the

annual production of the different technologies in the start year.

switch_peaking_capacity ::Int16 TODO.

set_peaking_slack ::Float16 TODO.

set_peaking_minrun_share ::Float16 TODO.

set_peaking_res_cf ::Float16 TODO.

set_peaking_startyear ::Int16 Year in which the peaking constraint becomes active.

switch_peaking_with_storages ::Int16 Enables to fulfill the peaking constraint with storages.

switch_peaking_with_trade ::Int16 Enables to fulfill the peaking constraint with trade.

switch_peaking_minrun ::Int16 TODO.

switch_employment_calculation ::Int16 TODO.

switch_endogenous_employment ::Int16 TODO.

employment_data_file ::String TODO.

switch_dispatch ::Int8 Used to enable the dispatch run.

elmod_nthhour ::Int16 Step size in hour for the sampling in the time reduction algorithm.

The default is 0 since the preferred method is to define daystep and hourstep instead. It corresponds to 24*daystep + hourstep.

elmod_starthour ::Int16 Starting hour for the sampling in the time reduction algorithm.

elmod_dunkelflaute ::Int16 Enables the addition of a period with very low wind and

sun in the winter during the time reduction algorithm.

elmod_daystep ::Int16 Number of days between each sample during the time reduction algorithm.

elmod_hourstep ::Int16 Number of hours in addition tothe day step between each sample.

switch_raw_results ::Int8 Used to enable the writing of raw results after a model run.

The raw results dumps the content of all variables into CSVs.

switch_processed_results ::Int8 Used to produce processed result files containing

additional metrics not part of the raw results.

write_reduced_timeserie ::Int8 Used to enable the writing of a file containing the

results of the time reduction algorithm.

extr_str_results ::String Final name of the result files written by the model.

extr_str_dispatch ::String If switch_dispatch = 1, final name of the result file form the investment

run that will be read to fix some decision variables.

switch_reserve ::Int16 Used to enable reserve margin constraints

GENeSYS_MOD.Tags — Type

Tags - The different tags used in the model to map across sets and subsets

GENeSYS_MOD.Variable_Parameters — Type

Intermediary variables calculated after the model run

The intermediary variables are used to ex-post aggregate the activity, demand and use by time, technology, mod of operation and/or region.

Methods

GENeSYS_MOD._registered_variables — Method

Return all variables in the model

GENeSYS_MOD.convert_jump_container_to_df — Method

Returns a DataFrame with the values of the variables from the JuMP container var. The column names of the DataFrame can be specified for the indexing columns in dim_names, and the name of the data value column by a Symbol value_col e.g. :Value

GENeSYS_MOD.create_daa — Method

Creates DenseAxisArrays containing the input parameters to the model considering hierarchy with base region data and world data.

The function creates a DenseAxisArray for a given parameter indexed by the given sets. The values are intialized to 0. If copy world is true, the value for the region world are copied. If inheritbaseworld is 1, missing data will be fetched from the base region if they exist and again from the world region if necessary.

GENeSYS_MOD.create_daa_hourly — Method

Internal function used in the run process after reading the input data to reduce the hourly timeseries for the whole year to a given number of timeslices. The algorithm maintain the max and min value and also fit the new timeserie to minimise deviation from the mean of the original timeseire.

GENeSYS_MOD.create_daa_init — Function

Create dense axis array initialized at a given value.

GENeSYS_MOD.create_df_hourly — Method

GENeSYS_MOD.genesysmod_bounds — Method

Internal function used in the run process to modify batches of input data.

GENeSYS_MOD.genesysmod_dataload — Method

Internal function used in the run process to load the input data and create the reduced timeseries.

GENeSYS_MOD.genesysmod_dec — Method

Internal function used in the run process to define the model variables.

GENeSYS_MOD.genesysmod_emissionintensity — Method

Internal function used in the run to compute sectoral emissions and emission intensity of fuels.

GENeSYS_MOD.genesysmod_employment — Method

GENeSYS_MOD.genesysmod_equ — Method

Internal function used in the run process to define the model constraints.

GENeSYS_MOD.genesysmod_levelizedcosts — Method

Internal function used in the run process to compute results related to levelized costs.

GENeSYS_MOD.genesysmod_results — Method

Internal function used in the run process to compute results. It also runs the functions for processing emissions and levelized costs.

GENeSYS_MOD.genesysmod_results_raw — Method

Write the values of each variable in the model to CSV files.

GENeSYS_MOD.genesysmod_settings — Method

Internal function used in the run process to set run settings such as dicount rates.

GENeSYS_MOD.genesysmod_variable_parameter — Method

Internal function used in the run process after solving to compute aggregated versions of the rate of activity, rate of use and demand, on mode of operation, timeslice and technology.

GENeSYS_MOD.make_mapping — Method

make_mapping(Sets,Params)

Creates a mapping of the allowed combinations of technology and fuel (and revers) and mode of operations.

GENeSYS_MOD.print_iis — Method

Write a text file containing the iis.

The function is used to write the iis to a file. By default the file is written in the working directory and is named iis.txt. The function compute_conflict!(model) must be run beforehands. The iis contains the set of constraint causing the infeasibility.

GENeSYS_MOD.simplify_iis — Method

Helper function to simplify the iis file and make the analysis easier.

The function simplify the iis by finding the equation of type x = 0 and reoplacing x in the other equations instead of doing that by hand. The round_numerics and digits optional parameters can also be used to reduce the length of equation by rounding numbers. It allows a faster analysis of iis files.

GENeSYS_MOD.split_by_substrings — Method

Extension of split function to take lists of substrings and keep delimiters

GENeSYS_MOD.timeseries_reduction! — Method