Public interface

Index

GENeSYSMOD.CSVResult

GENeSYSMOD.NoInfeasibilityTechs

GENeSYSMOD.NoRawResult

GENeSYSMOD.OneNodeSimple

GENeSYSMOD.OneNodeStorage

GENeSYSMOD.TXTResult

GENeSYSMOD.TXTandCSV

GENeSYSMOD.TwoNodes

GENeSYSMOD.WithInfeasibilityTechs

GENeSYSMOD.genesysmod

GENeSYSMOD.genesysmod_build_model

GENeSYSMOD.genesysmod_build_model_dispatch

GENeSYSMOD.genesysmod_dispatch

Types

GENeSYSMOD.CSVResult — Type

CSVResult - Raw results will be printed in CSV format. Each variable will have its CSV file named after the variable.

GENeSYSMOD.NoInfeasibilityTechs — Type

NoInfeasibilityTechs - type to use when the infeasibility techs are not needed.

GENeSYSMOD.NoRawResult — Type

NoRawResult - The raw results will not be printed.

GENeSYSMOD.OneNodeSimple — Type

OneNodeSimple - The dispatch is conducted for the considered_region in isolation from bordering regions.

GENeSYSMOD.OneNodeStorage — Type

OneNodeStorage - The dispatch is conducted for the considered_region while the interaction with the rest of the system is modeled as storages.

GENeSYSMOD.TXTResult — Type

TXTResult - Raw results will be printed in TXT format. A single txt file containing all values will be printed. This type as an argument filename used to define the name to use for the txt file. By default the argument takes the value "results".

GENeSYSMOD.TXTandCSV — Type

TXTandCSV - Raw results will be printed in both TXT and CSV formats.

GENeSYSMOD.TwoNodes — Type

TwoNodes - The dispatch is conducted for the considered_region and another dummy region consisting of an aggregation of the other regions.

GENeSYSMOD.WithInfeasibilityTechs — Type

WithInfeasibilityTechs - type to use when infeasibility techs are desired in the model. Typically they are used to help troubleshoot infeasibility as they provide a relaxation that allow the model to solve and to find the origin of theinfeasibility in the results.

Methods

GENeSYSMOD.genesysmod — Method

Run the whole model. It runs the whole process from input data reading to result processing. For information about the switches, refer to the datastructure documentation.

GENeSYSMOD.genesysmod_build_model — Method

Build the model without running it. For information about the switches, refer to the datastructure documentation.

GENeSYSMOD.genesysmod_build_model_dispatch — Method

Build the dispatch model. A previous run is necessary to allow to read in investment decisions. For information about the switches, refer to the datastructure documentation

GENeSYSMOD.genesysmod_dispatch — Method

Run the simple dispatch model. A previous run is necessary to allow to read in investment decisions. For information about the switches, refer to the datastructure documentation