Public interface

Index

GENeSYS_MOD.CSVResult

GENeSYS_MOD.NoInfeasibilityTechs

GENeSYS_MOD.NoRawResult

GENeSYS_MOD.OneNodeSimple

GENeSYS_MOD.OneNodeStorage

GENeSYS_MOD.TXTResult

GENeSYS_MOD.TXTandCSV

GENeSYS_MOD.TwoNodes

GENeSYS_MOD.WithInfeasibilityTechs

GENeSYS_MOD.genesysmod

GENeSYS_MOD.genesysmod_dispatch

Types

GENeSYS_MOD.CSVResult — Type

CSVResult - Raw results will be printed in CSV format. Each variable will have its CSV file named after the variable.

GENeSYS_MOD.NoInfeasibilityTechs — Type

NoInfeasibilityTechs - type to use when the infeasibility techs are not needed.

GENeSYS_MOD.NoRawResult — Type

NoRawResult - The raw results will not be printed.

GENeSYS_MOD.OneNodeSimple — Type

OneNodeSimple - The dispatch is conducted for the considered_region in isolation from bordering regions.

GENeSYS_MOD.OneNodeStorage — Type

OneNodeStorage - The dispatch is conducted for the considered_region while the interaction with the rest of the system is modeled as storages.

GENeSYS_MOD.TXTResult — Type

TXTResult - Raw results will be printed in TXT format. A single txt file containing all values will be printed. This type as an argument filename used to define the name to use for the txt file. By default the argument takes the value "results".

GENeSYS_MOD.TXTandCSV — Type

TXTandCSV - Raw results will be printed in both TXT and CSV formats.

GENeSYS_MOD.TwoNodes — Type

TwoNodes - The dispatch is conducted for the considered_region and another dummy region consisting of an aggregation of the other regions.

GENeSYS_MOD.WithInfeasibilityTechs — Type

WithInfeasibilityTechs - type to use when infeasibility techs are desired in the model. Typically they are used to help troubleshoot infeasibility as they provide a relaxation that allow the model to solve and to find the origin of theinfeasibility in the results.

Methods

GENeSYS_MOD.genesysmod — Method

Run the whole model. It runs the whole process from input data reading to result processing. For information about the switches, refer to the datastructure documentation.

GENeSYS_MOD.genesysmod_dispatch — Method

Run the simple dispatch model. A previous run is necessary to allow to read in investment decisions. For information about the switches, refer to the datastructure documentation