Public interface

Index

GENeSYS_MOD.genesysmod

GENeSYS_MOD.genesysmod_simple_dispatch

Types

Methods

GENeSYS_MOD.genesysmod — Method

Run the whole model. It runs the whole process from input data reading to result processing. For information about the switches, refer to the datastructure documentation.

GENeSYS_MOD.genesysmod_simple_dispatch — Method

Run the simple dispatch model. A previous run is necessary to allow to read in investment decisions. For information about the switches, refer to the datastructure documentation